fanfou-android: add udpate version page, replace popwindow icons, replace writepage icons, modify some classes, remove unused resources.

+++ androidapp/src/com/fanfou/app/ui/widget/TouchImageView.java
@@ -147,18 +147,18 @@ public class TouchImageView extends ImageView implements View.OnTouchListener {
 		height = MeasureSpec.getSize(heightMeasureSpec);
 		// Fit to screen.
 		float scale;
-		float scaleX = (float) width / (float) bmWidth;
-		float scaleY = (float) height / (float) bmHeight;
+		float scaleX = width / bmWidth;
+		float scaleY = height / bmHeight;
 		scale = Math.min(scaleX, scaleY);
 		matrix.setScale(scale, scale);
 		setImageMatrix(matrix);
 		saveScale = 1f;
 
 		// Center the image
-		redundantYSpace = (float) height - (scale * (float) bmHeight);
-		redundantXSpace = (float) width - (scale * (float) bmWidth);
-		redundantYSpace /= (float) 2;
-		redundantXSpace /= (float) 2;
+		redundantYSpace = height - (scale * bmHeight);
+		redundantXSpace = width - (scale * bmWidth);
+		redundantYSpace /= 2;
+		redundantXSpace /= 2;
 
 		matrix.postTranslate(redundantXSpace, redundantYSpace);
